requirements.txt
Django>=3.2,<5.0
djangorestframework>=3.12,<4.0
djangorestframework-simplejwt>=5.0,<6.0
